Separate studentkortet tracking
